Bump to v0.1.5:
This could be that I found a new way of blogging.
This could be about "Learning the ropes releasing Ruby gems".
And maybe about: why the hell do you bump releases to fix your ChangeLog
while you could actually use an experimental branch for this, and keep
the Git log clean and functional, for us engineers. Are you wasting our
precious time? I hope not. I hope that this will help others not go
through the same mistakes that I barely saw documented. Including the
actual functionality of the gem.

Cheers!